Extract Hack functionality to FromHHVM source
